Weight units (#3)
* Added new units

* Refactoring

* Added missing addition tests

* Added substraction tests

* Changed notation

* Added science notation to tests

* Removed unnecessary units

* Revert "Removed unnecessary units"

This reverts commit de1ff6f4de4616ebbcd48b28ac569bc6d10df133.

* Fixed science notation

* Changed factor to big decimal and introduced science notation

* Optimized constructors

* Fixed typo

* Added toBigDecimal method

* Imported members to simplify tests

* Removed to big decimal and added 2 units tests subtraction

* Added missing substraction weight tests

* Added missing subtraction tests to length

* Updated readme and version

* Readme update

+++ measurement/src/main/java/pl/dawidfiruzek/measurement/units/Units.kt
@@ -1,6 +1,8 @@
 package pl.dawidfiruzek.measurement.units
 
+import java.math.BigDecimal
+
 interface Units {
-    val factor: Double
+    val factor: BigDecimal
     fun getBaseUnits(): Units
 }
